backend/requirements.txt
fastapi
uvicorn
sqlalchemy
reportlab
python-multipart
google-generativeai
